🐛 fix cnspec build
