+++ apps/cnspec/cmd/scan.go
@@ -3,14 +3,9 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"os"
 	"sort"
 	"strings"
-
-	"go.mondoo.com/cnquery/motor/discovery/common"
-	"go.mondoo.com/cnspec"
-	"go.mondoo.com/ranger-rpc/plugins/scope"
 
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
@@ -24,6 +19,7 @@ import (
 	"go.mondoo.com/cnquery/cli/inventoryloader"
 	"go.mondoo.com/cnquery/cli/sysinfo"
 	"go.mondoo.com/cnquery/motor/asset"
+	"go.mondoo.com/cnquery/motor/discovery/common"
 	v1 "go.mondoo.com/cnquery/motor/inventory/v1"
 	"go.mondoo.com/cnquery/motor/providers"
 	"go.mondoo.com/cnquery/resources"
@@ -504,38 +500,6 @@ func getCobraScanConfig(cmd *cobra.Command, args []string, provider providers.Pr
 	}
 
 	return &conf, nil
-}
-
-func defaultRangerPlugins(sysInfo *sysinfo.SystemInfo, features cnquery.Features) []ranger.ClientPlugin {
-	plugins := []ranger.ClientPlugin{}
-	plugins = append(plugins, scope.NewRequestIDRangerPlugin())
-	plugins = append(plugins, sysInfoHeader(sysInfo, features))
-	return plugins
-}
-
-func sysInfoHeader(sysInfo *sysinfo.SystemInfo, features cnquery.Features) ranger.ClientPlugin {
-	const (
-		HttpHeaderUserAgent      = "User-Client"
-		HttpHeaderClientFeatures = "Mondoo-Features"
-		HttpHeaderPlatformID     = "Mondoo-PlatformID"
-	)
-
-	h := http.Header{}
-	info := map[string]string{
-		"cnspec": cnspec.Version,
-		"build":  cnspec.Build,
-	}
-	if sysInfo != nil {
-		info["PN"] = sysInfo.Platform.Name
-		info["PR"] = sysInfo.Platform.Version
-		info["PA"] = sysInfo.Platform.Arch
-		info["IP"] = sysInfo.IP
-		info["HN"] = sysInfo.Hostname
-		h.Set(HttpHeaderPlatformID, sysInfo.PlatformId)
-	}
-	h.Set(HttpHeaderUserAgent, scope.XInfoHeader(info))
-	h.Set(HttpHeaderClientFeatures, features.Encode())
-	return scope.NewCustomHeaderRangerPlugin(h)
 }
 
 func (c *scanConfig) loadPolicies() error {
